+++ src/cubey/vulkan/rendering.cpp
@@ -44,6 +44,71 @@ ImageLayoutTransition begin_depth_attachment_transition(VkImage image) {
     };
 }
 
+ImageLayoutTransition begin_storage_image_write_transition(VkImage image) {
+    return {
+        .image = image,
+        .aspect_mask = VK_IMAGE_ASPECT_COLOR_BIT,
+        .old_layout = VK_IMAGE_LAYOUT_UNDEFINED,
+        .new_layout = VK_IMAGE_LAYOUT_GENERAL,
+        .src_access_mask = 0,
+        .dst_access_mask = VK_ACCESS_SHADER_WRITE_BIT,
+        .src_stage_mask = VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
+        .dst_stage_mask = VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT,
+    };
+}
+
+ImageLayoutTransition finish_storage_image_write_for_sampling_transition(VkImage image) {
+    return {
+        .image = image,
+        .aspect_mask = VK_IMAGE_ASPECT_COLOR_BIT,
+        .old_layout = VK_IMAGE_LAYOUT_GENERAL,
+        .new_layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL,
+        .src_access_mask = VK_ACCESS_SHADER_WRITE_BIT,
+        .dst_access_mask = VK_ACCESS_SHADER_READ_BIT,
+        .src_stage_mask = VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT,
+        .dst_stage_mask = VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT,
+    };
+}
+
+ImageLayoutTransition begin_transfer_dst_transition(VkImage image) {
+    return {
+        .image = image,
+        .aspect_mask = VK_IMAGE_ASPECT_COLOR_BIT,
+        .old_layout = VK_IMAGE_LAYOUT_UNDEFINED,
+        .new_layout = VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL,
+        .src_access_mask = 0,
+        .dst_access_mask = VK_ACCESS_TRANSFER_WRITE_BIT,
+        .src_stage_mask = VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT,
+        .dst_stage_mask = VK_PIPELINE_STAGE_TRANSFER_BIT,
+    };
+}
+
+ImageLayoutTransition finish_transfer_dst_for_sampling_transition(VkImage image) {
+    return {
+        .image = image,
+        .aspect_mask = VK_IMAGE_ASPECT_COLOR_BIT,
+        .old_layout = VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL,
+        .new_layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL,
+        .src_access_mask = VK_ACCESS_TRANSFER_WRITE_BIT,
+        .dst_access_mask = VK_ACCESS_SHADER_READ_BIT,
+        .src_stage_mask = VK_PIPELINE_STAGE_TRANSFER_BIT,
+        .dst_stage_mask = VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT,
+    };
+}
+
+ImageLayoutTransition begin_transfer_src_transition(VkImage image) {
+    return {
+        .image = image,
+        .aspect_mask = VK_IMAGE_ASPECT_COLOR_BIT,
+        .old_layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL,
+        .new_layout = VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL,
+        .src_access_mask = VK_ACCESS_SHADER_READ_BIT,
+        .dst_access_mask = VK_ACCESS_TRANSFER_READ_BIT,
+        .src_stage_mask = VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT,
+        .dst_stage_mask = VK_PIPELINE_STAGE_TRANSFER_BIT,
+    };
+}
+
 VkImageMemoryBarrier image_memory_barrier(const ImageLayoutTransition& transition) {
     auto barrier = vk_struct<VkImageMemoryBarrier>(VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER);
     barrier.oldLayout = transition.old_layout;